+++ demisto_sdk/commands/common/tests/integration_test.py
@@ -55,6 +55,31 @@ class TestIntegrationValidator:
         validator = IntegrationValidator(structure)
         assert validator.is_added_required_fields() is answer
 
+    IS_REMOVED_INTEGRATION_PARAMETERS_INPUTS = [
+        ({"configuration": [{"name": "test"}]}, {"configuration": [{"name": "test"}]}, False),
+        ({"configuration": [{"name": "test"}, {"name": "test2"}]}, {"configuration": [{"name": "test"}]}, False),
+        ({"configuration": [{"name": "test"}]}, {"configuration": [{"name": "test"}, {"name": "test2"}]}, True),
+        ({"configuration": [{"name": "test"}]}, {"configuration": [{"name": "old_param"}, {"name": "test2"}]}, True),
+    ]
+
+    @pytest.mark.parametrize("current_file, old_file, answer", IS_REMOVED_INTEGRATION_PARAMETERS_INPUTS)
+    def test_is_removed_integration_parameters(self, current_file, old_file, answer):
+        """
+        Given
+        - integration configuration with different parameters
+
+        When
+        - running the validation is_removed_integration_parameters()
+
+        Then
+        - upon removal of parameters: it should set is_valid to False and return True
+        - upon non removal or addition of parameters: it should set is_valid to True and return False
+        """
+        structure = mock_structure("", current_file, old_file)
+        validator = IntegrationValidator(structure)
+        assert validator.is_removed_integration_parameters() is answer
+        assert validator.is_valid is not answer
+
     CONFIGURATION_JSON_1 = {"configuration": [{"name": "test", "required": False}, {"name": "test1", "required": True}]}
     EXPECTED_JSON_1 = {"test": False, "test1": True}
     FIELD_TO_REQUIRED_INPUTS = [
@@ -132,7 +157,7 @@ class TestIntegrationValidator:
     def test_with_duplicate_params(self, print_error):
         """
         Given
-        - integration configuratiton contains duplicate parameter (called test)
+        - integration configuration contains duplicate parameter (called test)
 
         When
         - running the validation is_there_duplicate_params()