+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Fixed an issue where running **validate**, **update-release-notes** outside of content repo crushed without a meaningful error message.
 * Added support for layoutscontainer in **init** contribution flow.
 * Added a validation for tlp_color param in feeds in **validate** command.
+* Added a validation for removal of integration parameters in **validate** command.
 * Fixed an issue where **update-release-notes** was failing with a wrong error message when no pack or input was given.
 
 # 1.2.1